requirements.txt
click==6.6
cloudant==2.4.0
ordereddict==1.1
pysolr==3.6.0
python-dotenv==0.5.1
requests==2.11.0
six==1.10.0
slackclient==1.0.5
watson-developer-cloud==0.25.1
websocket-client==0.37.0
